Correos P. Tablas y Vistas

+++ app/Http/Controllers/PedidoController.php
@@ -5,18 +5,21 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Pedido;
 use App\Models\User;
+use App\Http\Controllers\Controller;
 
 class PedidoController extends Controller
 {
     public function index()
     {
-        $pedidos = Pedido::select('id', 'id_usuario', 'total', 'created_at', 'updated_at')->with('user')->orderBy('updated_at', 'asc')->orderBy('id_usuario', 'asc')->get();
+        $this->authorize('isAdmin');
+        $pedidos = Pedido::select('id', 'id_usuario', 'total', 'created_at')->with('user')->orderBy('created_at', 'asc')->orderBy('id_usuario', 'asc')->get();
         return view("pedidos.pedidos_index", compact('pedidos'));
     }
 
 
     public function show($usuario_id) //Index User
     {
+        $this->authorize('isAdmin');
         $usuario = User::find($usuario_id);
         $pedidos_info = $usuario->pedidos;
         $nombre = $usuario->name;
@@ -26,8 +29,11 @@ class PedidoController extends Controller
 
     public function destroy($pedido_id)
     {
+        $this->authorize('isAdmin');
         $pedido = Pedido::find($pedido_id);
         $pedido->delete();
-        return redirect()->route('pedidos.index')->with('success', 'Pedido eliminado exitosamente.');
+
+        session()->flash('success');
+        return redirect()->route('pedidos.index')->with('success', 'Pedido eliminado exitosamente');
     }
 }
