+++ app/Http/Controllers/ProductoController.php
@@ -4,18 +4,24 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Producto;
-use App\Models\Pedido;
+//use Illuminate\Support\Facades\Gate;
+use App\Http\Controllers\Controller;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Support\Facades\Storage;
 
 class ProductoController extends Controller
 {
+
     public function index()
     {
-        $productos = Producto::select('id', 'nombre', 'precio', 'created_at', 'updated_at')->orderBy('nombre', 'asc')->get();
+        $this->authorize('isAdmin');
+        $productos = Producto::select('id', 'nombre', 'precio', 'descripcion', 'photo_path', 'created_at', 'updated_at')->orderBy('id', 'asc')->get();
         return view("productos.productos_index", compact('productos'));
     }
 
     public function create()
     {
+        $this->authorize('isAdmin');
         return view("productos.productos_create");
     }
 
@@ -24,14 +30,31 @@ class ProductoController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('isAdmin');
         $request->validate([
-            'nombre' => 'required|string|max:255',
-            'precio' => 'required|numeric|min:5|max:500',
+            'nombre' => 'required|string|min:5|max:255',
+            'precio' => 'required|numeric|between:5.00,500.00',
+            'descripcion' => 'required|string|min:5|max:255',
+            'photo_path' => 'required|image'
         ]);
 
-        Producto::create($request->all());
+        if ($request->hasFile('photo_path')) {
+            $name = time() . '.' . $request->file('photo_path')->getClientOriginalExtension();
+            $img = $request->file('photo_path')->storeAs('public/img', $name);
+            $imagePath = '/img/' . $name;
+        }
 
-        return redirect()->route("productos.index");
+        $producto = new Producto([
+            'nombre' => $request->input('nombre'),
+            'precio' => $request->input('precio'),
+            'descripcion' => $request->input('descripcion'),
+            'photo_path' => $imagePath ?? null,
+        ]);
+
+        $producto->save();
+
+        session()->flash('success');
+        return redirect()->route("productos.index")->with('success', 'Producto creado exitosamente');
     }
 
     /**
@@ -39,6 +62,7 @@ class ProductoController extends Controller
      */
     public function show($id) //Index Producto
     {
+        $this->authorize('isAdmin');
         $pro = Producto::find($id);
         return view('productos.productos_show', compact('pro'));
     }
@@ -48,6 +72,7 @@ class ProductoController extends Controller
      */
     public function edit($producto_id)
     {
+        $this->authorize('isAdmin');
         $producto = Producto::find($producto_id);
         return view("productos.productos_edit", compact('producto'));
     }
@@ -55,28 +80,50 @@ class ProductoController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, $producto)
+    public function update(Request $request, $producto_id)
     {
+        $this->authorize('isAdmin');
+
+        $producto = Producto::find($producto_id);
+
         $request->validate([
-            'nombre' => 'required|string|max:255',
-            'precio' => 'required|numeric|min:5|max:500',
+            'nombre' => 'required|string|min:5|max:255',
+            'precio' => 'required|numeric|between:5,500',
+            'descripcion' => 'required|string|min:5|max:255',
+            'photo_path' => 'nullable|image'
         ]);
 
-        $producto = Producto::find($producto);
-        $producto->nombre = $request->input('nombre');
-        $producto->precio = $request->input('precio');
-        $producto->save();
+        if ($request->hasFile('photo_path')) {
+            // Elimina la imagen anterior si existe
+            Storage::disk('public')->delete($producto->photo_path);
+            $name = time() . '.' . $request->file('photo_path')->getClientOriginalExtension();
+            $img = $request->file('photo_path')->storeAs('public/img', $name);
+            $imagePath = '/img/' . $name;
+            $producto->photo_path = $imagePath;
+        }
 
-        return redirect()->route("productos.index");
+        $producto->update([
+            'nombre' => $request->input('nombre'),
+            'precio' => $request->input('precio'),
+            'descripcion' => $request->input('descripcion'),
+        ]);
+
+        session()->flash('success');
+        return redirect()->route("productos.index")->with('success', 'Producto actualizado exitosamente');
     }
+
+
 
     /**
      * Remove the specified resource from storage.
      */
     public function destroy($producto_id)
     {
+        $this->authorize('isAdmin');
         $producto = Producto::find($producto_id);
         $producto->delete();
-        return redirect()->route('productos.index');
+
+        session()->flash('success');
+        return redirect()->route('productos.index')->with('success', 'Producto eliminado exitosamente');
     }
 }
